Implementa experimento ASR disponibilidad: API Gateway con Heartbeat y Active Redundancy para detección y recuperación de fallos en < 1 segundo

+++ monitor_servicios/models.py
@@ -7,7 +7,7 @@ class ServiceRegistration(models.Model):
     """
     name = models.CharField(max_length=100, unique=True)
     description = models.TextField(blank=True)
-    expected_interval_seconds = models.PositiveIntegerField(default=30)
+    expected_interval_seconds = models.FloatField(default=30.0)
     is_active = models.BooleanField(default=True)
     registered_at = models.DateTimeField(auto_now_add=True)
